# grov-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Grōv Restoration Platform API",
        "description": "Production-quality REST API backend for Grōv field restoration tracking, monitoring, community hub, and leaderboard.",
        "contact": {
            "name": "Grōv Engineering Team",
            "email": "support@grov.app"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "Grōv Primary API Server v1"
        }
    ],
    "paths": {
        "/activities/my-activities": {
            "get": {
                "tags": [
                    "Activities"
                ],
                "summary": "Get user's logged activities",
                "description": "Returns list of current user's activities with filter options (all, plantation, seeding, monitored).",
                "operationId": "dfc6844eb032df1344400fddb49601d8",
                "parameters": [
                    {
                        "name": "filter",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "all",
                                "plantation",
                                "seeding",
                                "monitored"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Activities retrieved successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/activities/{id}": {
            "get": {
                "tags": [
                    "Activities"
                ],
                "summary": "Get single activity details",
                "description": "Returns full activity information, stepper timeline, photos, and monitoring history.",
                "operationId": "941e12e56ef66240b71dfe799a602dc0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Activity details retrieved successfully"
                    },
                    "404": {
                        "description": "Activity not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/activities/plantation": {
            "post": {
                "tags": [
                    "Activities"
                ],
                "summary": "Log a tree plantation activity",
                "description": "Logs sapling planting, creates location, stores photo evidence, awards server-side points.",
                "operationId": "ded34b15ed6922a6a51d98833a214f57",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "species_id",
                                    "quantity_planted",
                                    "date",
                                    "site_name",
                                    "latitude",
                                    "longitude",
                                    "planting_method"
                                ],
                                "properties": {
                                    "species_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "quantity_planted": {
                                        "type": "integer",
                                        "example": 120
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-08-27"
                                    },
                                    "site_name": {
                                        "type": "string",
                                        "example": "Margalla Hills Sector 3"
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 33.7294
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 73.0931
                                    },
                                    "planting_method": {
                                        "type": "string",
                                        "enum": [
                                            "Pit Planting",
                                            "Trench Planting",
                                            "Mound Planting",
                                            "Aerial Planting"
                                        ]
                                    },
                                    "field_notes": {
                                        "type": "string"
                                    },
                                    "photos[]": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Plantation activity logged successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/activities/seeding": {
            "post": {
                "tags": [
                    "Activities"
                ],
                "summary": "Log a seed bombing activity",
                "description": "Logs seed dispersal, creates location, stores photo evidence, awards server-side points.",
                "operationId": "2991cc7054f140aadf1840496212d5c8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "species_id",
                                    "seeds_dispersed",
                                    "date",
                                    "site_name",
                                    "latitude",
                                    "longitude",
                                    "dispersal_method"
                                ],
                                "properties": {
                                    "species_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "seeds_dispersed": {
                                        "type": "integer",
                                        "example": 500
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-08-27"
                                    },
                                    "site_name": {
                                        "type": "string",
                                        "example": "G-11 Margalla Trail"
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 33.7294
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 73.0931
                                    },
                                    "dispersal_method": {
                                        "type": "string",
                                        "enum": [
                                            "Hand Broadcasting",
                                            "Seed Bombing (aerial)",
                                            "Seed Drill",
                                            "Hydroseeding"
                                        ]
                                    },
                                    "coverage_area_sqm": {
                                        "type": "integer",
                                        "example": 2000
                                    },
                                    "field_notes": {
                                        "type": "string"
                                    },
                                    "photos[]": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Seeding activity logged successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter Sanctum bearer token in format: Bearer {token}",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}